+++ src/BankInterface.java
@@ -8,11 +8,11 @@ public interface BankInterface extends Remote {
     
 	public long login(String username, String password) throws RemoteException, InvalidLogin;//, InvalidLogin;
 	
-	public void deposit(int accountnum, int amount, long sessionID) throws RemoteException, InvalidSession; //, InvalidSession;
+	public void deposit(int accountnum, float amount, long sessionID) throws RemoteException, InvalidSession; //, InvalidSession;
 	
-	public void withdraw(int accountnum, int amount, long sessionID) throws RemoteException, InvalidSession;
+	public void withdraw(int accountnum, float amount, long sessionID) throws RemoteException, InvalidSession;
 	
-	public int inquiry(int accountnum, long sessionID) throws RemoteException, InvalidSession;
+	public float inquiry(int accountnum, long sessionID) throws RemoteException, InvalidSession;
 	
 	public ArrayList<Transaction> getStatement(int accountnum, Date from, Date to, long sessionID) throws RemoteException, InvalidSession;
